Update the delete and update user

+++ app/Http/Controllers/API/UserController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UserRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
@@ -32,27 +33,46 @@ class UserController extends Controller
 
     public function update(UserRequest $request, $id)
     {
-        $user = User::findOrFail($id);
+        $status = 'success';
+        $message = 'User Successfully Updated';
 
-        $user->name = $request->name;
-        $user->email = $request->email;
-        $user->user_type = $request->user_type;
-        $user->save();
+        if ($id == '1' && Auth::id() == '1') {
+            $status = 'warning';
+            $message = 'Cannot update your account';
+        } else {
+            $user = User::findOrFail($id);
+    
+            $user->name = $request->name;
+            $user->email = $request->email;
+            $user->user_type = $request->user_type;
+            $user->save();
+        }
 
         return response()->json([
-            'status' => 'success',
-            'message' => 'User Successfully Updated'
+            'status' => $status,
+            'message' => $message
         ], Response::HTTP_OK);
     }
 
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
+        $status = 'success';
+        $message = 'User Successfully Deleted';
+
+        if (Auth::id() == $id) {
+            $status = 'warning';
+            $message = 'Cannot delete your account';
+        } else if ($id == '1') {
+            $status = 'warning';
+            $message = 'Cannot delete Super Admin';
+        } else {
+            $user = User::findOrFail($id);
+            $user->delete();
+        }
 
         return response()->json([
-            'status' => 'success',
-            'message' => 'User Successfully Deleted'
+            'status' => $status,
+            'message' => $message
         ], Response::HTTP_OK);
     }
 }
